[CMake] add PYTHON_EXT_SUFFIX back to the main library
This re-apply https://github.com/stack-of-tasks/eigenpy/pull/57, because
the main library is linked to boost python, and in case we have a setup
with multiple python versions, we want to ensure that the python module
will not link to a version of the main library linked to the wrong boost
python library.

This is a reccuring issue for users who are using Python 3 on Ubuntu
18.04 or 20.04, while they also have ROS, which is stuck to a Python 2
version of eigenpy.
eg. https://github.com/stack-of-tasks/eigenpy/issues/230

This was removed in https://github.com/stack-of-tasks/eigenpy/pull/113.

Instead of deactivating this feature for everybody because this can
raise an issue on Windows (I have no idea what this issue is), this
commit reactivate this feature for everybody except Windows users.